requirements.txt
selenium
requests
beautifulsoup4
anytree
Pillow
imagehash